# new-site/packages/backend/docs/swagger.yaml
basePath: /
definitions:
...
  user.ResendVerificationRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  user.SafeUser:
    properties:
      age:
        type: integer
      city:
        type: string
      disabilities:
        type: string
      education:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      gender:
        type: string
      hasPapfe:
        type: boolean
      linkedin:
        type: string
      name:
        type: string
      presence_rate:
        type: number
      profession:
        type: string
      telegram:
        type: string
      user_number:
        type: string
    type: object
  user.UpdateUserRequest:
    properties:
      age:
        type: integer
      city:
        type: string
      disabilities:
        type: string
      education:
        type: string
      email:
        type: string
      gender:
        type: string
      hasPapfe:
        type: boolean
      linkedin:
        type: string
      name:
        type: string
      profession:
        type: string
      telegram:
        type: string
    required:
    - age
    - city
    - disabilities
    - education
    - email
    - gender
    - name
    type: object
  user.VerifyEmailRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
...
host: localhost:4000
info:
  contact: {}
  description: API do backend da Semcomp.
  title: Semcomp API
  version: "1.0"
paths:
...
  /admin/users:
    get:
      consumes:
      - application/json
      description: Retorna uma lista paginada de usuários cadastrados
      parameters:
      - default: 1
        description: Página atual
        in: query
        name: page
        type: integer
      - default: 10
        description: Limite de itens por página
        in: query
        name: limit
        type: integer
      - default: name
        description: Campo de ordenação
        in: query
        name: sort_by
        type: string
      - default: asc
        description: Ordem (asc/desc)
        in: query
        name: sort_order
        type: string
      - description: Campo de busca
        in: query
        name: search_by
        type: string
      - description: Valor de busca
        in: query
        name: search_value
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista de usuários paginada
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Parâmetro inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Lista usuários
      tags:
      - Usuários (Participantes)
  /admin/users/{id}:
    delete:
      consumes:
      - application/json
      description: Remove um usuário do sistema
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Usuário removido com sucesso
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: ID inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Usuário não existe
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Deleta usuário
      tags:
      - Usuários (Participantes)
    get:
      consumes:
      - application/json
      description: Retorna os dados de um usuário específico
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Usuário encontrado
          schema:
            $ref: '#/definitions/user.SafeUser'
        "400":
          description: ID inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Usuário não encontrado
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Busca usuário por ID
      tags:
      - Usuários (Participantes)
    put:
      consumes:
      - application/json
      description: Altera os dados de um usuário existente
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: integer
      - description: Dados para atualização
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Usuário atualizado com sucesso
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Dados ou ID inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Email já cadastrado
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Atualiza usuário
      tags:
      - Usuários (Participantes)
...
  /register:
    post:
      consumes:
      - multipart/form-data
      description: Cadastra um participante no sistema. Aceita multipart/form-data
        para suportar upload do comprovante PAPFE.
      parameters:
      - description: Nome completo
        in: formData
        name: name
        required: true
        type: string
      - description: E-mail
        in: formData
        name: email
        required: true
        type: string
      - description: Senha (min 8 caracteres)
        in: formData
        name: password
        required: true
        type: string
      - description: Idade
        in: formData
        name: age
        required: true
        type: integer
      - description: Gênero
        in: formData
        name: gender
        required: true
        type: string
      - description: Cidade
        in: formData
        name: city
        required: true
        type: string
      - description: Formação
        in: formData
        name: education
        required: true
        type: string
      - description: Recebe PAPFE?
        in: formData
        name: hasPapfe
        type: boolean
      - collectionFormat: csv
        description: Lista de deficiências
        in: formData
        items:
          type: string
        name: disabilities
        type: array
      - description: Profissão
        in: formData
        name: profession
        type: string
      - description: LinkedIn
        in: formData
        name: linkedin
        type: string
      - description: Telegram
        in: formData
        name: telegram
        type: string
      - description: Comprovante PAPFE (PDF/JPEG/PNG/WebP, máx 10MB)
        in: formData
        name: papfe_document
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Usuário criado com sucesso!
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Dados inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: E-mail já cadastrado
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cria um novo usuário
      tags:
      - Usuários (Participantes)
  /resend-verification:
    post:
      consumes:
      - application/json
      description: Reenvia o link de confirmação de e-mail, se aplicável
      parameters:
      - description: E-mail do usuário
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.ResendVerificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Se o e-mail estiver cadastrado e pendente de confirmação, um
            novo link foi enviado
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Dados inválidos
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reenvia e-mail de confirmação
      tags:
      - Usuários (Participantes)
  /verify-email:
    post:
      consumes:
      - application/json
      description: Valida o token de verificação enviado por e-mail e ativa a conta
      parameters:
      - description: Token de verificação
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: E-mail confirmado com sucesso!
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Dados inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Token inválido ou expirado
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Confirma e-mail de um usuário
      tags:
      - Usuários (Participantes)
securityDefinitions:
  BearerAuth:
    description: 'Insira o token no formato: Bearer {seu_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"